Made semantic refactoring
