+++ README.md
@@ -1,0 +1,7 @@
+IngridientItem забыли типизировать
+Нужно посмотреть все компоненты и типизировать их пропсы, если где-то пропустили.
+Всем элементам пропсов, которые являются обязательными, нужно добавлять isRequired при типизации с propTypes. Это касается самих массивов и объектов. В самом конце нужно указывать isRequired, если массив сам обязателен в props
+Почитайте внимательно задание 1го спринта. Модальное окно должно выглядеть не так. И должны быть обработчики Escape и оверлея внутри
+внутри компонента Modal не должно быть хардкода строк и компонентов. IngredientDetails и OrderDetails нужно передавать в качестве children.
+Внутри компонента Modal и ModalOverlay не стоит использовать dispatch для закрытия попапов, так как для каждого отдельного окна обычно разная логика закрытия (и в следующих спринтах как раз это и будет). Вам нужно передавать функцию закрытия onClose в props модальному окну, как это было раньше
+Modal является универсальным компонентом для любых модальных окон. Он не должен знать внутри себя какие-то конкретные данные (экшены) для конкретных попапов.
