+++ src/components/app-header/app-header.jsx
@@ -8,32 +8,34 @@ import styles from './app-header.module.css';
 
 export const AppHeader = () => {
   return (
-    <div className={styles.wrapper}>
-      <div className={styles.header}>
-        <div className={styles.info}>
-          <div className="mt-4 mb-4">
-            <div className={styles.constructor}>
+    <header>
+      <nav>
+        <ul className={styles.info}>
+          <li className="mt-4 mb-4">
+            <a href="#">
               <BurgerIcon />
-              <div>Конструткор</div>
-            </div>
-          </div>
-          <div className="mt-4 mb-4">
-            <div className={styles.orders}>
+              <span>Конструткор</span>
+            </a>
+          </li>
+          <li className="mt-4 mb-4">
+            <a href="#">
               <ListIcon type="secondary" />
-              <div>Лента заказов</div>
-            </div>
-          </div>
-        </div>
-        <div className="logo">
+              <span>Лента заказов</span>
+            </a>
+          </li>
+        </ul>
+        <ul>
           <Logo />
-        </div>
-        <div className="mt-4 mb-4">
-          <div className={styles.cabinet}>
-            <ProfileIcon />
-            <div>Личный кабинет</div>
-          </div>
-        </div>
-      </div>
-    </div>
+        </ul>
+        <ul>
+          <li className={styles.cabinet}>
+            <a href="#">
+              <ProfileIcon />
+              <span>Личный кабинет</span>
+            </a>
+          </li>
+        </ul>
+      </nav>
+    </header>
   );
 };
