+++ src/components/app-header/app-header.module.css
@@ -1,9 +1,9 @@
-.wrapper {
+header {
   width: 100%;
   background-color: #1C1C21;
 }
 
-.header {
+nav {
   max-width: 1280px;
   margin: 0 auto;
   display: flex;
@@ -13,16 +13,25 @@
   height: 88px;
 }
 
+header nav ul {
+  list-style: none;
+  padding: 0;
+}
+
+header a {
+  text-decoration: none;
+  color: white;
+  display: flex;
+  align-items: center;
+  column-gap: 8px;
+  padding: 0 20px;
+}
+
 .info {
   display: flex;
   column-gap: 8px;
 }
 
-.constructor,
-.orders,
 .cabinet {
-  cursor: pointer;
-  display: flex;
-  padding: 16px 20px;
-  column-gap: 8px;
+  margin-left: 120px;
 }
