+++ src/components/app/app.module.css
@@ -13,7 +13,7 @@ code {
     monospace;
 }
 
-.main {
+main {
   max-width: 1280px;
   margin: 0 auto;
   display: flex;
