+++ src/index.tsx
@@ -1,22 +1,16 @@
 import ReactDOM from 'react-dom/client';
-import { App } from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import './index.css';
 import store from './redux/store';
+import {App} from './components/app/app'
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
- 
   <Provider store={store}>
-   <App />
+    <App/>
   </Provider>
-
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
